requirements.txt
# Core Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0

# Database
mysql-connector-python==8.1.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
xlrd==2.0.1

# Authentication and Security
bcrypt==4.0.1

# File Transfer
paramiko==3.3.1

# Data Validation and Analysis
numexpr==2.8.5

# Environment Management
python-dotenv==1.0.0

# Utilities
secrets==1.0.0
logging==0.4.9.6

# Date and Time
python-dateutil==2.8.2

# Text Processing
regex==2023.8.8

# Math and Statistics
scipy==1.11.2
scikit-learn==1.3.0

# Caching (optional)
redis==4.6.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development Tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Production Server
gunicorn==21.2.0
